style: estilizando popups do index
